% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randoms.R
\name{posteriorSummary}
\alias{posteriorSummary}
\title{posteriorSummary function
I still need to document all of these things properly}
\usage{
posteriorSummary(paramSampleVec)
}
\description{
posteriorSummary function
I still need to document all of these things properly
}
